adopt the workflow task vocabulary from next major
