+++ facefusion/workflows/image_to_image.py
@@ -1,25 +1,26 @@
 from functools import partial
 
-from facefusion import ffmpeg
-from facefusion import logger, process_manager, state_manager, translator
+from facefusion import content_analyser, ffmpeg, logger, process_manager, state_manager, translator
 from facefusion.audio import create_empty_audio_frame
-from facefusion.content_analyser import analyse_image
 from facefusion.filesystem import is_image
 from facefusion.processors.core import get_processors_modules
-from facefusion.temp_helper import clear_temp_directory, create_temp_directory, get_temp_file_path
+from facefusion.temp_helper import get_temp_file_path
 from facefusion.time_helper import calculate_end_time
 from facefusion.types import ErrorCode
 from facefusion.vision import conditional_merge_vision_mask, detect_image_resolution, extract_vision_mask, pack_resolution, read_static_image, read_static_images, restrict_image_resolution, scale_resolution, write_image
-from facefusion.workflows.core import is_process_stopping
+from facefusion.workflows.core import clear, is_process_stopping, setup
 
 
 def process(start_time : float) -> ErrorCode:
 	tasks =\
 	[
+		analyse_image,
+		clear,
 		setup,
 		prepare_image,
 		process_image,
-		partial(finalize_image, start_time)
+		partial(finalize_image, start_time),
+		clear
 	]
 	process_manager.start()
 
@@ -34,16 +35,9 @@ def process(start_time : float) -> ErrorCode:
 	return 0
 
 
-def setup() -> ErrorCode:
-	if analyse_image(state_manager.get_item('target_path')):
+def analyse_image() -> ErrorCode:
+	if content_analyser.analyse_image(state_manager.get_item('target_path')):
 		return 3
-
-	if clear_temp_directory(state_manager.get_item('target_path')):
-		logger.debug(translator.get('clearing_temp'), __name__)
-
-	if create_temp_directory(state_manager.get_item('target_path')):
-		logger.debug(translator.get('creating_temp'), __name__)
-
 	return 0
 
 
@@ -103,9 +97,6 @@ def finalize_image(start_time : float) -> ErrorCode:
 		logger.debug(translator.get('finalizing_image_succeeded'), __name__)
 	else:
 		logger.warn(translator.get('finalizing_image_skipped'), __name__)
-
-	logger.debug(translator.get('clearing_temp'), __name__)
-	clear_temp_directory(state_manager.get_item('target_path'))
 
 	if is_image(state_manager.get_item('output_path')):
 		logger.info(translator.get('processing_image_succeeded').format(seconds = calculate_end_time(start_time)), __name__)
